src/modules/accounts/useCases/createUser/CreateUserController.ts
import { Request, Response } from "express";
import { container } from "tsyringe";

import { CreateUserUseCase } from "./CreateUserUseCase";
import { AppError } from "@shared/errors/AppError";

class CreateUserController {
  async handle(request: Request, response: Response): Promise<Response> {
    const { username, password } = request.body;

    if (!username || !password) {
      throw new AppError("Please provide username and password", 401);
    }
    
    const createUserUseCase = container.resolve(CreateUserUseCase);

    await createUserUseCase.execute({
      username,
      password,
    });

    return response.status(201).send();
  }
}

export { CreateUserController };
